S15.f.7: extract plugin.ts handoff orchestration (5598 → 5071 LOC)
Seventh sub-slice of S15.f.

cli/src/core/agt-handoff.ts (new, 609 LOC):
  runHandoffOrchestration(token, adminToken, direction, dirLabel, deps)
    — multi-step agent state transfer (cloud<->local) over the mesh:
    encrypted state snapshot, sub-agent state collection,
    re-registration, ratchet-to-successor.
  _hp (now a closure inside, capturing progress + log)
  HandoffProgress, AgtInboxEntry, HandoffDeps interfaces.

HandoffDeps threads the four state pieces the function touches:
  - progress (HandoffProgress, by reference — mutations propagate)
  - inbox (AgtInboxEntry[], by reference — splice/findIndex)
  - meshClient/identity accessors
  - meshSend (forwards to plugin.ts wrapper capturing agtIdentity)

Body byte-identical apart from agtMeshClient -> deps.meshClient()
and agtIdentity -> deps.identity(); nameToAmid/amidToName imported
directly from core/amid-cache.ts.

Cumulative S15.f delta:
  pre:  7139
  f.1:  6974  (-165)
  f.2:  6890   (-84)
  f.3:  6648  (-242)
  f.4:  6488  (-160)
  f.5:  6104  (-384)
  f.6:  5598  (-506)
  f.7:  5071  (-527)
  Cumulative: -2068. 2071 LOC remaining to §4.2 cap of 3000 (69% there).

Verification:
  - tsc --noEmit clean / lint 30 (unchanged) / build clean
  - npm test: 454 pass / 2 skipped (baseline)

Audit: docs/security-audits/2026-04-29-phase2-hotspot-plugin-cli-f7.md
